pyproject.toml
[tool.poetry]
name = "project-1"
version = "0.1.0"
description = ""
authors = ["Alexey Ryzhkov <aler32@mail.ru>"]
readme = "README.md"
packages = [{include = "project_1"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.1"
pillow = "^9.5.0"
drf-yasg = "^1.21.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
